Cosmetics: suppress a compiler warning.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@77549 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/util/test/test_row_reader_performance.cpp
@@ -96,7 +96,7 @@ int CRowReaderPerfTest::Run(void)
         for (Int8 item = 0; item < item_count; ++item) {
             if (item != 0)
                 fputc(' ', f);
-            fprintf(f, "%ld", item);
+            fprintf(f, "%ld", (long)(item));
         }
     }
     fclose(f);
